split stories to separate files
